move class-proj-rel ofProject API from lb3 to lb4
