+++ client/projects/lib-sdk-lb4/src/lib/sdk-lb4/api/proClassFieldConfig.service.ts
@@ -85,92 +85,6 @@ export class ProClassFieldConfigService {
     }
 
     /**
-     * Creates or updates instances of ProClassFieldConfig.
-     * @param pkProject Pk of the project
-     * @param proClassFieldConfig Array ProClassFieldConfig
-     * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
-     * @param reportProgress flag to report request and response progress.
-     */
-    public proClassFieldConfigBulkUpsert(pkProject: number, proClassFieldConfig: Array<ProClassFieldConfig>, observe?: 'body', reportProgress?: boolean, options?: {httpHeaderAccept?: 'application/json' | 'application/xml' | 'text/xml' | 'application/javascript' | 'text/javascript'}): Observable<ProClassFieldConfig>;
-    public proClassFieldConfigBulkUpsert(pkProject: number, proClassFieldConfig: Array<ProClassFieldConfig>, observe?: 'response', reportProgress?: boolean, options?: {httpHeaderAccept?: 'application/json' | 'application/xml' | 'text/xml' | 'application/javascript' | 'text/javascript'}): Observable<HttpResponse<ProClassFieldConfig>>;
-    public proClassFieldConfigBulkUpsert(pkProject: number, proClassFieldConfig: Array<ProClassFieldConfig>, observe?: 'events', reportProgress?: boolean, options?: {httpHeaderAccept?: 'application/json' | 'application/xml' | 'text/xml' | 'application/javascript' | 'text/javascript'}): Observable<HttpEvent<ProClassFieldConfig>>;
-    public proClassFieldConfigBulkUpsert(pkProject: number, proClassFieldConfig: Array<ProClassFieldConfig>, observe: any = 'body', reportProgress: boolean = false, options?: {httpHeaderAccept?: 'application/json' | 'application/xml' | 'text/xml' | 'application/javascript' | 'text/javascript'}): Observable<any> {
-        if (pkProject === null || pkProject === undefined) {
-            throw new Error('Required parameter pkProject was null or undefined when calling proClassFieldConfigBulkUpsert.');
-        }
-        if (proClassFieldConfig === null || proClassFieldConfig === undefined) {
-            throw new Error('Required parameter proClassFieldConfig was null or undefined when calling proClassFieldConfigBulkUpsert.');
-        }
-
-        let queryParameters = new HttpParams({encoder: this.encoder});
-        if (pkProject !== undefined && pkProject !== null) {
-          queryParameters = this.addToHttpParams(queryParameters,
-            <any>pkProject, 'pkProject');
-        }
-
-        let headers = this.defaultHeaders;
-
-        let credential: string | undefined;
-        // authentication (accesstoken) required
-        credential = this.configuration.lookupCredential('accesstoken');
-        if (credential) {
-            headers = headers.set('authorization', credential);
-        }
-
-        // authentication (jwt) required
-        credential = this.configuration.lookupCredential('jwt');
-        if (credential) {
-            headers = headers.set('Authorization', 'Bearer ' + credential);
-        }
-
-        let httpHeaderAcceptSelected: string | undefined = options && options.httpHeaderAccept;
-        if (httpHeaderAcceptSelected === undefined) {
-            // to determine the Accept header
-            const httpHeaderAccepts: string[] = [
-                'application/json',
-                'application/xml',
-                'text/xml',
-                'application/javascript',
-                'text/javascript'
-            ];
-            httpHeaderAcceptSelected = this.configuration.selectHeaderAccept(httpHeaderAccepts);
-        }
-        if (httpHeaderAcceptSelected !== undefined) {
-            headers = headers.set('Accept', httpHeaderAcceptSelected);
-        }
-
-
-        // to determine the Content-Type header
-        const consumes: string[] = [
-            'application/json',
-            'application/x-www-form-urlencoded',
-            'application/xml',
-            'text/xml'
-        ];
-        const httpContentTypeSelected: string | undefined = this.configuration.selectHeaderContentType(consumes);
-        if (httpContentTypeSelected !== undefined) {
-            headers = headers.set('Content-Type', httpContentTypeSelected);
-        }
-
-        let responseType: 'text' | 'json' = 'json';
-        if(httpHeaderAcceptSelected && httpHeaderAcceptSelected.startsWith('text')) {
-            responseType = 'text';
-        }
-
-        return this.httpClient.put<ProClassFieldConfig>(`${this.configuration.basePath}/lb3-api/ProClassFieldConfigs/bulk-upsert`,
-            proClassFieldConfig,
-            {
-                params: queryParameters,
-                responseType: <any>responseType,
-                withCredentials: this.configuration.withCredentials,
-                headers: headers,
-                observe: observe,
-                reportProgress: reportProgress
-            }
-        );
-    }
-
-    /**
      * Patch an existing model instance or insert a new one into the data source.
      * @param proClassFieldConfig Model instance data
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
