+++ server/.env.example
@@ -41,4 +41,5 @@ MOCHA_FOLDER=**/warehouse/**
 
 # ONTOME settings
 
-ONTOME_URL=https://ontome.net # the base url of ontome
+# the base url of ontome
+ONTOME_URL=https://ontome.net
