+++ server/lb3app/common/models/pro-dfh-class-proj-rel.json
@@ -48,7 +48,6 @@
       "permission": "ALLOW",
       "property": [
         "create",
-        "ofProject",
         "bulkUpsert"
       ]
     }
